wake lock: Remove WakeLockPermissionDescriptor, use separate permission names.
This adapts the implementation to some of the breaking spec changes
introduced in https://github.com/w3c/screen-wake-lock/pull/255.

Namely, WakeLockPermissionDescriptor has been removed in favor of using a
regular PermissionDescriptor object with |name| set to "screen-wake-lock".

That spec change also removed system wake locks (and consequently the
additions to the WorkerNavigator interface) altogether, but that will be
done in a separate CL, so for now we also support a PermissionDescriptor
object with |name| set to "system-wake-lock".

Bug: 257511, 1064685
Change-Id: I096f3fae24444da5ca6b1eeb3889a813a9e5f270