[css-filters] Migrate background-image-blur-repaint.html test to WPT
Migrate this test out of third_party/blink/web_tests/css3/filters
and into the WPT-specific directory, adding links to the relevant
specs and a test assertion describing its purpose.

Note that this test depends on the runAfterLayoutAndPaint() helper
function from run-after-layout-and-paint.js, which is based on
testRunner. Thus, this migration also includes the required changes
to the test to implement the same functionality in the reftest,
that is, relying on two chained runs of requestAnimationFrame()
to make sure that at least one frame has been rendered and then
using <html class="reftest-wait"> as described in [1] to make
sure we don't take the screenshot until the desired changes have
been made (i.e. shrinking a box, in this particular test).

Last, a new file in external/wpt/css/filter-effects/support has
been added as required by this test, as well as by many other ones
under css3/filters that will be migrated in the future.

[1] https://web-platform-tests.org/writing-tests/reftests.html?highlight=reftests#controlling-when-comparison-occurs

Bug: 1063749
Change-Id: Iebc5db044597853ab47d723337c6235a84f9cf5b