[css-filters] Migrate direct-image-dynamic-filter.html test to WPT
Migrate this test out of third_party/blink/web_tests/css3/filters
and into the WPT-specific directory, adding links to the relevant
specs and a test assertion describing its purpose.

In addition, this test uses <html class="reftest-wait"> as described
in [1] to replace the code that was depending on testRunner in order
to make sure that the screenshot is taking at the right time.

Last, the original test relied on a green-256x256.jpg image that is
exactly equivalent to the PNG version already in WPT, so this test
has been migrated to use that one instead.

[1] https://web-platform-tests.org/writing-tests/reftests.html?highlight=reftests#controlling-when-comparison-occurs

Bug: 1063749
Change-Id: Idcd76fcf3b1e82fe14be602c120d8c105f5c658a