[css-filters] Migrate effect-drop-shadow-clip-abspos.html test to WPT
Migrate this test out of third_party/blink/web_tests/css3/filters
and into the WPT-specific directory, adding links to the relevant
specs and a test assertion describing its purpose.

Since there was already a similar test already in WPT, this CL
also renames it adding the -001 suffix to it, so that we can
consistently add this new one as a -002 one.

Bug: 1063749
Change-Id: Ibb6f0ef3783cf9805e339f70b0e7616369cbc752