Handle k-rate AudioParam inputs for BiquadFilterNode
In addition to making BiquadFilterNode AudioParams apply the input to
the param, we also needed to fix an issue in
CalculateSampleAccurateValues when the param is k-rate.  We would sum
in all the values for the input instead of just taking one.

Manually ran the new test case with Chrome stable and all the tests fail
as expected.

Bug: 1015760
Test: k-rate-biquad-connections.html

Change-Id: I412f9567909d8dd6d4cf00036138b47a454c431c